remove cycle between flycam_accessor and view_mode_accessor.ts
